refactor(sequence): separate residue policy and validation reports
