+++ packages/rust/biors-core/tests/core_contract.rs
@@ -3,7 +3,8 @@ use biors_core::{
     parse_fasta_records_reader, protein_20_vocabulary, stable_input_hash, tokenize_fasta_records,
     tokenize_fasta_records_reader, validate_fasta_input, validate_fasta_reader, BioRsError,
     FastaReadError, ModelInputBuildError, ModelInputPolicy, PaddingPolicy, ProteinSequence,
-    ProteinTokenizer, Tokenizer, UnknownTokenPolicy, PROTEIN_20_UNKNOWN_TOKEN_ID,
+    ProteinTokenizer, ResidueIssue, Tokenizer, UnknownTokenPolicy, ValidatedSequence,
+    PROTEIN_20_UNKNOWN_TOKEN_ID,
 };
 use std::io::{Cursor, ErrorKind};
 use std::path::Path;
@@ -70,6 +71,53 @@ fn validates_sequences_with_ambiguous_residue_policy() {
     assert_eq!(report.records, 1);
     assert_eq!(report.valid_records, 0);
     assert_eq!(report.warning_count, 1);
+    assert_eq!(report.error_count, 1);
+}
+
+#[test]
+fn sequence_validation_report_aggregates_record_validity_and_issue_counts() {
+    let report = biors_core::sequence::summarize_validated_sequences(vec![
+        ValidatedSequence {
+            id: "valid".to_string(),
+            sequence: "ACDE".to_string(),
+            alphabet: "protein-20".to_string(),
+            valid: true,
+            warnings: vec![],
+            errors: vec![],
+        },
+        ValidatedSequence {
+            id: "warn".to_string(),
+            sequence: "XB".to_string(),
+            alphabet: "protein-20".to_string(),
+            valid: false,
+            warnings: vec![
+                ResidueIssue {
+                    residue: 'X',
+                    position: 1,
+                },
+                ResidueIssue {
+                    residue: 'B',
+                    position: 2,
+                },
+            ],
+            errors: vec![],
+        },
+        ValidatedSequence {
+            id: "invalid".to_string(),
+            sequence: "A?".to_string(),
+            alphabet: "protein-20".to_string(),
+            valid: false,
+            warnings: vec![],
+            errors: vec![ResidueIssue {
+                residue: '?',
+                position: 2,
+            }],
+        },
+    ]);
+
+    assert_eq!(report.records, 3);
+    assert_eq!(report.valid_records, 1);
+    assert_eq!(report.warning_count, 2);
     assert_eq!(report.error_count, 1);
 }
